requirements.txt
pandas~=1.3.1
numpy~=1.21.2
matplotlib~=3.4.2
pillow~=8.3.1
seaborn~=0.11.1
scikit-learn~=0.24.2
requests~=2.26.0
bs4~=4.9.3
beautifulsoup4~=4.9.3
yaml~=0.2.5
pyyaml~=5.4.1
pytest~=6.2.4
tensorflow~=2.3.0
tensorflow-gpu~=2.3.0
opencv-python~=4.5.4.58